Refactor  user managment section
