+++ server/src/components/Invitations/InvitationsTable.tsx
@@ -1,4 +1,3 @@
-import { MoreVertical } from "lucide-react";
 "use client";
 
 import {
